Update patch set 5
Patch Set 5:

> If you are going to set a flag to false here that is set to true in
 > msm8916-common, please reconsider. That's usually the sign of
 > something wrong, or bad commonization.

Hi Ethan,

I talked to arcee about MALLOC_IMPL. He said it is set to dlmalloc because of some memory corruption/alignment issues in 64-bit. He is oaky with using jemalloc for rendang (as long as there is not memory issue). Since Rendang was shipped with dlmalloc disabled, I think it make sense to keep it the same.

arcee [12:35] 
jemalloc is more efficient
arcee [12:36]
so if we don't have problems from using it, then go ahead
arcee [12:36]
but be sure :simple_smile:

Thanks,
Lin

Patch-set: 5