Update patch set 5
Patch Set 5:

> That's not what I'm talking about. I'm saying that you shouldn't
 > set a flag in a device-common/BoardConfigCommon.mk, then unset it
 > in the device/BoardConfig.mk. Just move the flag down to all
 > device/BoardConfig.mk and remove it from BoardConfigCommon.mk. (If
 > that is in fact what is happening)

I'm not sure if arcee is on board with that. I can take the MALLOC_IMPL flag out from the common mk and add in device trees of these that need it, if every one is okay with that. For now maybe I should remove it from this patch...

Patch-set: 5